% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_estado_bancos.R
\name{cmf_besb_data}
\alias{cmf_besb_data}
\title{Importar archivo estado resultado}
\usage{
cmf_besb_data(url_file)
}
\arguments{
\item{url_file}{url a descargar y leer.}
}
\description{
Importar archivo estado resultado
}
\examples{

cmf_besb_data("https://www.cmfchile.cl/portal/estadisticas/617/articles-43001_recurso_1.zip")

}
