% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_estado_bancos.R
\name{cmf_besb_importar}
\alias{cmf_besb_importar}
\title{Importar archivo estado resultado}
\usage{
cmf_besb_importar(archivo)
}
\arguments{
\item{archivo}{archivo a importar.}
}
\description{
Importar archivo estado resultado
}
